# api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MediAsk API",
    "version": "1.0.0",
    "description": "智能医疗辅助问诊系统 - 按模块组织"
  },
  "servers": [
    {
      "url": "http://localhost:8989",
      "description": "开发环境"
    }
  ],
...
  "paths": {
...
    "/api/v1/admin/authz/permissions": {
      "get": {
        "tags": [
          "authz"
        ],
        "summary": "查询权限列表",
        "description": "返回系统内全部权限编码定义",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthzPermissionListResultResponse"
                }
              }
            }
          }
        },
        "operationId": "getApiV1AdminAuthzPermissions"
      }
    },
    "/api/v1/admin/authz/roles": {
      "get": {
        "tags": [
          "authz"
        ],
        "summary": "查询角色列表",
        "description": "返回角色基础信息及绑定的权限编码",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthzRoleListResultResponse"
                }
              }
            }
          }
        },
        "operationId": "getApiV1AdminAuthzRoles"
      }
    },
    "/api/v1/admin/authz/users/{userId}/roles": {
      "get": {
        "tags": [
          "authz"
        ],
        "summary": "查询用户角色",
        "description": "返回指定用户当前角色编码列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "用户ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRolesResultResponse"
                }
              }
            }
          }
        },
        "operationId": "getApiV1AdminAuthzUsersByUserIdRoles"
      },
      "put": {
        "tags": [
          "authz"
        ],
        "summary": "更新用户角色",
        "description": "覆盖式更新用户角色编码列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "用户ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRolesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "putApiV1AdminAuthzUsersByUserIdRoles"
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "用户登录",
        "description": "支持用户名或手机号登录，响应 data 内含 token、refreshToken、refreshTokenId、expiresIn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功，响应 data 为登录信息（含 token/refreshToken/refreshTokenId/expiresIn）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResultResponse"
                }
              }
            }
          }
        },
        "security": [],
        "operationId": "postApiV1AuthLogin"
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "用户登出",
        "description": "撤销当前访问令牌所属用户的 Refresh Token 会话。当前实现不接收请求体。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "登出成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "postApiV1AuthLogout"
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "刷新Token",
        "description": "使用 refreshToken 换取新的 access token，并轮换 refreshToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "刷新成功，响应 data 为新的登录信息（含 token/refreshToken/refreshTokenId/expiresIn）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResultResponse"
                }
              }
            }
          }
        },
        "security": [],
        "operationId": "postApiV1AuthRefresh"
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "用户注册",
        "description": "注册成功后自动登录，响应 data 内含 token、refreshToken、refreshTokenId、expiresIn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功，响应 data 为登录信息（含 token/refreshToken/refreshTokenId/expiresIn）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResultResponse"
                }
              }
            }
          }
        },
        "security": [],
        "operationId": "postApiV1AuthRegister"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AuthzPermissionListResultResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "状态码，0表示成功"
          },
          "msg": {
            "type": "string",
            "description": "提示信息"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionResponse"
            }
          }
        }
      },
      "AuthzRoleListResultResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "状态码，0表示成功"
          },
          "msg": {
            "type": "string",
            "description": "提示信息"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleResponse"
            }
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "account",
          "password"
        ],
        "properties": {
          "account": {
            "type": "string",
            "description": "用户名或手机号"
          },
          "password": {
            "type": "string",
            "description": "密码"
          }
        },
        "example": {
          "account": "test_user",
          "password": "password123"
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64",
            "description": "用户ID"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "userType": {
            "type": "integer",
            "format": "int32",
            "description": "用户类型"
          },
          "authorities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "权限列表"
          },
          "tokenType": {
            "type": "string",
            "description": "Token 类型",
            "example": "Bearer"
          },
          "token": {
            "type": "string",
            "description": "Access Token (30分钟有效)"
          },
          "expireAt": {
            "type": "integer",
            "format": "int64",
            "description": "Access Token 过期时间（秒级时间戳）"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int64",
            "description": "Access Token 剩余有效期（秒）"
          },
          "refreshToken": {
            "type": "string",
            "description": "Refresh Token (30天有效)"
          },
          "refreshTokenId": {
            "type": "string",
            "description": "Refresh Token ID（用于登出时撤销）"
          }
        },
        "example": {
          "userId": 1,
          "username": "admin",
          "userType": 3,
          "authorities": [
            "schedule:create",
            "schedule:auto",
            "schedule:update"
          ],
          "tokenType": "Bearer",
          "token": "eyJhbGciOiJIUzI1NiJ9...",
          "expireAt": 1700000000,
          "expiresIn": 1800,
          "refreshToken": "eyJhbGciOiJIUzI1NiJ9...",
          "refreshTokenId": "550e8400-e29b-41d4-a716-446655440000"
        }
      },
      "LoginResultResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "状态码，0表示成功"
          },
          "msg": {
            "type": "string",
            "description": "提示信息"
          },
          "data": {
            "$ref": "#/components/schemas/LoginResponse"
          }
        },
        "example": {
          "code": 0,
          "msg": "success",
          "data": {
            "userId": 1,
            "username": "admin",
            "userType": 3,
            "authorities": [
              "schedule:create",
              "schedule:auto",
              "schedule:update"
            ],
            "tokenType": "Bearer",
            "token": "eyJhbGciOiJIUzI1NiJ9...",
            "expireAt": 1700000000,
            "expiresIn": 1800,
            "refreshToken": "eyJhbGciOiJIUzI1NiJ9...",
            "refreshTokenId": "550e8400-e29b-41d4-a716-446655440000"
          }
        }
      },
...
      "PermissionResponse": {
        "type": "object",
        "properties": {
          "permissionId": {
            "type": "integer",
            "format": "int64",
            "description": "权限ID"
          },
          "permissionCode": {
            "type": "string",
            "description": "权限编码"
          },
          "permissionName": {
            "type": "string",
            "description": "权限名称"
          },
          "description": {
            "type": "string",
            "description": "权限描述"
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "required": [
          "refreshToken"
        ],
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "刷新令牌"
          }
        },
        "example": {
          "refreshToken": "eyJhbGciOiJIUzI1NiJ9..."
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": [
          "username",
          "password",
          "userType",
          "gender"
        ],
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "description": "密码（6-64位）"
          },
          "phone": {
            "type": "string",
            "description": "手机号（可选）"
          },
          "userType": {
            "type": "integer",
            "format": "int32",
            "description": "用户类型：1=患者，2=医生，3=管理员",
            "enum": [
              1,
              2,
              3
            ]
          },
          "realName": {
            "type": "string",
            "description": "真实姓名（可选）"
          },
          "gender": {
            "type": "integer",
            "format": "int32",
            "description": "性别：0=未知，1=男，2=女",
            "enum": [
              0,
              1,
              2
            ]
          },
          "birthDate": {
            "type": "string",
            "format": "date",
            "description": "出生日期（可选）"
          },
          "avatarUrl": {
            "type": "string",
            "description": "头像地址（可选）"
          }
        },
        "example": {
          "username": "test_user",
          "password": "password123",
          "phone": "13800138000",
          "userType": 3,
          "realName": "张三",
          "gender": 1,
          "birthDate": "1998-01-01",
          "avatarUrl": "https://example.com/avatar.png"
        }
      },
      "ResponseWrapper": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "状态码，0表示成功"
          },
          "msg": {
            "type": "string",
            "description": "提示信息"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "traceId": {
            "type": "string",
            "description": "链路追踪ID"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "响应时间戳（毫秒）"
          }
        },
        "example": {
          "code": 0,
          "msg": "success",
          "data": {}
        }
      },
      "RoleResponse": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "integer",
            "format": "int64",
            "description": "角色ID"
          },
          "roleCode": {
            "type": "string",
            "description": "角色编码"
          },
          "roleName": {
            "type": "string",
            "description": "角色名称"
          },
          "description": {
            "type": "string",
            "description": "角色描述"
          },
          "permissionCodes": {
            "type": "array",
            "description": "权限编码列表",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "UpdateUserRolesRequest": {
        "type": "object",
        "required": [
          "roleCodes"
        ],
        "properties": {
          "roleCodes": {
            "type": "array",
            "description": "角色编码列表（覆盖式更新）",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UserRolesResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64",
            "description": "用户ID"
          },
          "roleCodes": {
            "type": "array",
            "description": "角色编码列表",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UserRolesResultResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "状态码，0表示成功"
          },
          "msg": {
            "type": "string",
            "description": "提示信息"
          },
          "data": {
            "$ref": "#/components/schemas/UserRolesResponse"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}